refactor(web): rename landing-v2 to landing now that the rebrand is shipped
