+++ apps/web/app/blog/[slug]/page.tsx
@@ -5,7 +5,7 @@ import { PostMeta } from "../../components/blog/PostMeta"
 import { getAllPosts, getPost, getRelatedPosts } from "../../components/blog/post-index"
 import { DocsTOC } from "../../components/docs/DocsTOC"
 import { RelatedCards } from "../../components/docs/RelatedCards"
-import { FinalCta } from "../../components/landing-v2/FinalCta"
+import { FinalCta } from "../../components/landing/FinalCta"
 import { ReadingLayout } from "../../components/ReadingLayout"
 
 interface PageProps {
