+++ apps/web/app/components/landing/Hero.tsx
