+++ apps/web/app/components/landing/IntelliSenseVisual.tsx
