+++ apps/web/app/components/landing/ProofStrip.tsx
