+++ apps/web/app/page.tsx
@@ -1,15 +1,15 @@
-import { Ecosystem } from "./components/landing-v2/Ecosystem"
-import { Faq } from "./components/landing-v2/Faq"
-import { FeatureDevLoop } from "./components/landing-v2/FeatureDevLoop"
-import { FeatureRouting } from "./components/landing-v2/FeatureRouting"
-import { FeatureTools } from "./components/landing-v2/FeatureTools"
-import { FeatureTypes } from "./components/landing-v2/FeatureTypes"
-import { FinalCta } from "./components/landing-v2/FinalCta"
-import { Hero } from "./components/landing-v2/Hero"
-import { KeepTheRuntime } from "./components/landing-v2/KeepTheRuntime"
-import { ProofStrip } from "./components/landing-v2/ProofStrip"
-import { Quickstart } from "./components/landing-v2/Quickstart"
-import { WhyDawn } from "./components/landing-v2/WhyDawn"
+import { Ecosystem } from "./components/landing/Ecosystem"
+import { Faq } from "./components/landing/Faq"
+import { FeatureDevLoop } from "./components/landing/FeatureDevLoop"
+import { FeatureRouting } from "./components/landing/FeatureRouting"
+import { FeatureTools } from "./components/landing/FeatureTools"
+import { FeatureTypes } from "./components/landing/FeatureTypes"
+import { FinalCta } from "./components/landing/FinalCta"
+import { Hero } from "./components/landing/Hero"
+import { KeepTheRuntime } from "./components/landing/KeepTheRuntime"
+import { ProofStrip } from "./components/landing/ProofStrip"
+import { Quickstart } from "./components/landing/Quickstart"
+import { WhyDawn } from "./components/landing/WhyDawn"
 
 export default function HomePage() {
   return (
